+++ src/service/auth/SocialService.ts
@@ -1,6 +1,6 @@
 import { AxiosError } from 'axios';
 import HttpClient from '../../apis/HttpClient';
-import { IAuthProps, IAuthEmailProps, IAuthEmailVaildationProps } from './type';
+import { IAuthEmailProps, IAuthEmailVaildationProps } from './type';
 
 export const SocialService = async (code: string | null, state: string) => {
     try {
@@ -19,29 +19,6 @@ export async function SignInService(providerValue: string) {
     try {
         const res = await HttpClient.post(`/api/v1/auth/signin`, {
             providerId: `${providerValue}`,
-        });
-        return res.data;
-    } catch (error) {
-        console.error(error);
-    }
-}
-
-export async function SignUpService({
-    email,
-    categoryIds,
-    nickName,
-    provider,
-    providerId,
-    profileImageUrl,
-}: IAuthProps) {
-    try {
-        const res = await HttpClient.post(`/api/v1/auth/signup`, {
-            email: `${email}`,
-            categoryIds: categoryIds.map(Number),
-            nickname: `${nickName}`,
-            provider: `${provider}`,
-            providerId: `${providerId}`,
-            profileImageUrl: `${profileImageUrl}`,
         });
         return res.data;
     } catch (error) {
